Split openavb_endpoint_osal.c into multiple files
Split openavb_endpoint_osal.c into multiple files to make the code
easier to follow and modify.
Functionality should be unchanged.
